Release 2.0.1
Fixed some bugs that showed up in indigobuggie. Mostly to do with wb on
file open as python3 behaves differently from python3. It does mean they
will be issues with cross OS files as they won't merge do to the CR LF
issue with (windows, mac, and unix).

+++ beorn_lib/dialog/text_dialog/text_box_element.py
@@ -1,13 +1,13 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 #---------------------------------------------------------------------------------
-#                                                   
-#                    ,--.                                 
-#                    |  |-.  ,---.  ,---. ,--.--.,--,--,  
-#                    | .-. '| .-. :| .-. ||  .--'|      \ 
-#                    | `-' |\   --.' '-' '|  |   |  ||  | 
-#                     `---'  `----' `---' `--'   `--''--' 
-#                                                    
+#
+#                    ,--.
+#                    |  |-.  ,---.  ,---. ,--.--.,--,--,
+#                    | .-. '| .-. :| .-. ||  .--'|      \
+#                    | `-' |\   --.' '-' '|  |   |  ||  |
+#                     `---'  `----' `---' `--'   `--''--'
+#
 #    file: text_box
 #    desc: Text Box Element
 #
@@ -92,7 +92,7 @@ class TextBoxElement(BoxElement):
 			This function returns the cursor position relative to the start of
 			the screen.
 		"""
-		y_pos = self.y + ( self.text_pos / self.width)
+		y_pos = self.y + int( self.text_pos / self.width)
 		x_pos = self.x + len(self.title) + (self.text_pos % self.width) + 3
 		return (x_pos, y_pos)
