+++ beorn_lib/timekeeper/job.py
@@ -93,9 +93,9 @@ class Job(NestedTreeNode):
 		hour = minute * 60
 		day = hour * 24
 
-		days = self.total_time / day
-		hours = (self.total_time % day) / hour
-		minutes = (self.total_time % hour) / minute
+		days = int(self.total_time / day)
+		hours = int((self.total_time % day) / hour)
+		minutes = int((self.total_time % hour) / minute)
 
 		return "{:04d}:{:02d}:{:02d}".format(days, hours, minutes)
